+++ src/command_modules/azure-cli-botservice/HISTORY.rst
@@ -7,6 +7,7 @@ Release History
 * [Breaking Change]: The `--appid` and `--password` args for `az bot create` are now required and can be created via `az ad app create`.
 * [Breaking Change]: `az bot create -v v4` does not create or use a Storage Account or Application Insights.
 * [Breaking Change]: Instead of mapping Application Insights regions for `az bot create -v v3`, the command only accepts regions where Application Insights is creatable.
+* [Breaking Change]: `az bot update` is no longer a generic update command, but instead can affect specific properties of a bot.
 * [Deprecation]: All --lang flags now accept and advise users to use `Javascript` instead of `Node`. `Node` as a --lang value is deprecated.
 * Add `Typescript` support to `az bot prepare-deploy`
 * `az bot prepare-deploy` returns `true` if successful and has helpful verbose logging.